Module timely_communication::allocator::generic

A generic allocator, wrapping known implementors of Allocate.

This type is useful in settings where it is difficult to write code generic in A: Allocate, for example closures whose type arguments must be specified.

Enums

Generic
Enumerates known implementors of Allocate. Passes trait method calls on to members.