Struct timely_communication::allocator::zero_copy::bytes_slab::BytesSlab

pub struct BytesSlab { /* fields omitted */ }

A large binary allocation for writing and sharing.

A bytes slab wraps a Bytes and maintains a valid (written) length, and supports writing after this valid length, and extracting Bytes up to this valid length. Extracted bytes are enqueued and checked for uniqueness in order to recycle them (once all shared references are dropped).

Methods

impl BytesSlab

pub fn new(shift: usize) -> Self

Allocates a new BytesSlab with an initial size determined by a shift.

pub fn empty(&mut self) -> &mut [u8]

The empty region of the slab.

pub fn valid(&mut self) -> &mut [u8]

The valid region of the slab.

pub fn make_valid(&mut self, bytes: usize)

Marks the next bytes bytes as valid.

pub fn extract(&mut self, bytes: usize) -> Bytes

Extracts the first bytes valid bytes.

pub fn ensure_capacity(&mut self, capacity: usize)

Ensures that self.empty().len() is at least capacity.

This method may retire the current buffer if it does not have enough space, in which case it will copy any remaining contents into a new buffer. If this would not create enough free space, the shift is increased until it is sufficient.