Module timely_communication::initialize

Initialization logic for a generic instance of the Allocate channel allocation trait.

Enums

Configuration

Functions

initialize
Initializes an allocator::Generic for each thread, spawns the local threads, and invokes the supplied function with the allocator. Returns only once all threads have returned.