Module timely_communication::logging

Structs

ActiveBufferingLogger

BufferingLogger

CommsSetup

CommunicationEvent

SerializationEvent
Serialization

Enums

CommsEvent

LoggerBatch

Functions

get_precise_time_ns
Returns the value of an high resolution performance counter, in nanoseconds, rebased to be roughly comparable to an unix timestamp. Useful for comparing and merging logs from different machines (precision is limited by the precision of the wall clock base; clock skew effects should be taken into consideration).

Type Definitions

CommsLogger
A log writer for a communication thread.