Enum timely_communication::message::RefOrMut

pub enum RefOrMut<'a, T> where
    T: 'a,  {
    Ref(&'a T),
    Mut(&'a mut T),
}

Either an immutable or mutable reference.

Variants
Ref(&'a T)

An immutable reference.

Mut(&'a mut T)

A mutable reference.

Methods

impl<'a, T: Clone + 'a> RefOrMut<'a, T>

pub fn swap<'b>(self, element: &'b mut T)

Extracts the contents of self, either by cloning or swapping.

This consumes self because its contents are now in an unknown state.

pub fn replace(self, element: T) -> T

Extracts the contents of self, either by cloning or swapping.

This consumes self because its contents are now in an unknown state.