Struct timely_communication::message::Message

pub struct Message<T> { /* fields omitted */ }

A wrapped message which may be either typed or binary data.

Methods

impl<T> Message<T>

pub fn from_typed(typed: T) -> Self

Wrap a typed item as a message.

pub fn if_typed(self) -> Option<T>

Destructures and returns any typed data.

pub fn if_mut(&mut self) -> Option<&mut T>

Returns a mutable reference, if typed.

impl<T: Abomonation> Message<T>

pub unsafe fn from_bytes(bytes: Bytes) -> Self

Wrap bytes as a message.

#Safety

This method is unsafe, in that Abomonated::new() is unsafe: it presumes that the binary data can be safely decoded, which is unsafe for e.g. UTF8 data and enumerations (perhaps among many other types).

pub fn as_ref_or_mut(&mut self) -> RefOrMut<T>

Returns an immutable or mutable typed reference.

This method returns a mutable reference if the underlying data are typed Rust instances, which admit mutation, and it returns an immutable reference if the data are serialized binary data.

pub fn length_in_bytes(&self) -> usize

The number of bytes required to serialize the data.

pub fn into_bytes<W: Write>(&self, writer: &mut W)

Writes the binary representation into writer.

impl<T: Abomonation + Clone> Message<T>

pub fn into_typed(self) -> T

Produces a typed instance of the wrapped element.

pub fn as_mut(&mut self) -> &mut T

Ensures the message is typed data and returns a mutable reference to it.